+++ tests/test_query_rlzs.py
@@ -9,7 +9,7 @@ TOSHI_ID = 'FAk3T0sHi1D=='
 # vs30s = [250, 350, 450]
 imts = ['PGA', 'SA(0.5)']
 locs = ['WLG', 'QZN']
-rlzs = [f"rlz-00{x}" for x in range(5)]
+rlzs = [f"00{x}" for x in range(5)]
 lvps = list(map(lambda x: model.LevelValuePairAttribute(lvl=x / 1e3, val=(x / 1e6)), range(1, 51)))
 
 
@@ -62,12 +62,12 @@ class QueryRlzsTest(unittest.TestCase):
     def test_query_stats_objects_4(self):
         self.assertEqual(model.ToshiOpenquakeHazardCurveStats.exists(), True)
         query.batch_save_hcurve_rlzs(TOSHI_ID, models=build_rlzs_models())
-        res = list(query.get_hazard_rlz_curves(TOSHI_ID, ['PGA'], ['WLG', 'QZN'], ['rlz-001']))
+        res = list(query.get_hazard_rlz_curves(TOSHI_ID, ['PGA'], ['WLG', 'QZN'], ['001']))
         print(res)
         self.assertEqual(len(res), 2)
         self.assertEqual(res[0].loc, 'QZN')
         self.assertEqual(res[1].loc, 'WLG')
-        self.assertEqual(res[0].rlz, 'rlz-001')
+        self.assertEqual(res[0].rlz, '001')
 
     def test_query_stats_objects_all(self):
         self.assertEqual(model.ToshiOpenquakeHazardCurveStats.exists(), True)